api: update the use of as_ptr to ensure the reference to self is always valid
The existing use of as_ptr did not guarantee that self was always valid.  Depending on the optimizations of LLVM, the memory may have been freed immediately after creation rather than after the completion of the function call.  This left the pointee memory in a non-deterministic state causing garbage values to be send to the system calls.  Updated the usage of as_ptr to ensure the CString objects remained around through the lifetime of the pointer usage as described in std::ffi::CString as_ptr documentation.

Signed-off-by: Derek Straka <derek@asterius.io>